refactor: bookmark -> bookmarks 로 네이밍 수정
